requirements.txt
Django==4.0.5
djangorestframework==3.13.1
pandas==1.4.2
django-pandas~=0.6.6
